# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ChangePasswordRequest:
    properties:
      new_password:
        description: 新密码（至少6位）
        example: new123456
        type: string
      old_password:
        description: 原密码
        example: old123456
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  handler.DeleteMeRequest:
    properties:
      password:
        description: 当前密码确认
        example: secret123
        type: string
    required:
    - password
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: 'failed to read uploaded image: ...'
        type: string
    type: object
  handler.LoginRequest:
    properties:
      password:
        description: 密码
        example: admin123
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      expires_at:
        description: 过期时间（access 或 mfa_token 的）
        type: string
      mfa_required:
        description: MFARequired 为 true 时，需带上 mfa_token 调用 POST /auth/mfa/verify 完成登录。
        type: boolean
      mfa_token:
        description: MFAToken 是第二步验证用的短期令牌（5 分钟，一次性）。
        type: string
      token:
        description: 'JWT（放在 Authorization: Bearer <token> 头中）'
        type: string
      token_type:
        description: 恒为 Bearer
        type: string
      user:
        allOf:
        - $ref: '#/definitions/user.SafeUser'
        description: 用户信息（需要 MFA 第二步时不返回）
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      nickname:
        description: 昵称（可选）
        example: 张三
        type: string
      password:
        description: 密码（至少6位）
        example: secret123
        type: string
      username:
        description: 用户名（3-32位字母/数字/下划线）
        example: zhangsan
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  user.SafeUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      mfa_enabled:
        description: 是否已开启两步验证
        type: boolean
      nickname:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: wechatapp-web
    url: http://www.swagger.io/support
  description: |-
    Image background replacement API. Uploads an image, removes its
    background via the remove.bg service, then optionally composites
    the cutout onto a new background (solid color or another image).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: wechatapp-web API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Authenticates with username/password and returns a JWT to be
        sent as "Authorization: Bearer <token>".
      parameters:
      - description: Username and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT and user info
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Wrong username or password
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login and get a JWT
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the presented JWT so it can no longer be used.
      responses:
        "200":
          description: 'message: 已退出登录'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout (revoke current token)
      tags:
      - auth
  /auth/me:
    delete:
      consumes:
      - application/json
      description: |-
        Deletes the authenticated user's own account after verifying
        the password, and revokes the current token. The last admin
        cannot be deleted this way.
      parameters:
      - description: Password confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DeleteMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 账号已注销'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Wrong password, or last admin cannot be deleted
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User no longer exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete my own account (注销账号)
      tags:
      - auth
    get:
      description: Returns the profile of the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Current user profile
          schema:
            $ref: '#/definitions/user.SafeUser'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User no longer exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
...
  /auth/password:
    put:
      consumes:
      - application/json
      description: Changes the password of the authenticated user after verifying
        the old one.
      parameters:
      - description: Old and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 密码修改成功'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or wrong old password
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User no longer exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change own password
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a normal (user-role) account. Login afterwards to get a
        token.
      parameters:
      - description: Username, password and optional nickname
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/user.SafeUser'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Username already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT issued by POST /auth/login, sent as "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"